Add files via upload
First approach is the one I originally had which was brute-force that was incredibly slow thus making it inefficient to use. Brute-force has a time complexity of O(min(a,b)) 😭.
Second approach is using the Euclidean algorithm with a time complexity of O(log(min(a,b))) in which what I basically did was run a while loop till our variable is 0 and then swap the first variable's value with the second variable's value and replace the second variable's value with the divisor we get at that specific iteration and so in the end we will end up with the greatest divisor!